chore: move file + search only on enter
