+++ frontend/src/components/SearchBar/SearchBar.tsx
@@ -1,26 +1,31 @@
+import { useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 
+import { useSearch } from "../../context/SearchContext";
 import { Input } from "../retroui";
-import { useSearch } from "./SearchContext";
 import { cleanInput } from "./utils/cleanInput";
 
 export function SearchBar() {
   const navigate = useNavigate();
   const location = useLocation();
+
   const { query, setQuery } = useSearch();
+
+  const [inputValue, setInputValue] = useState(query);
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const cleaned = cleanInput(e.target.value);
-    setQuery(cleaned);
+    setInputValue(cleaned);
   };
 
   const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
     if (e.key === "Enter") {
       e.preventDefault();
+      setQuery(inputValue);
       if (location.pathname !== "/browse") {
-        navigate(`/browse?q=${encodeURIComponent(query)}`);
+        navigate(`/browse?q=${encodeURIComponent(inputValue)}`);
       } else {
-        navigate(`?q=${encodeURIComponent(query)}`, { replace: true });
+        navigate(`?q=${encodeURIComponent(inputValue)}`, { replace: true });
       }
     }
   };
@@ -28,7 +33,7 @@ export function SearchBar() {
   return (
     <Input
       type="text"
-      value={query}
+      value={inputValue}
       onChange={handleChange}
       onKeyDown={handleKeyDown}
       placeholder="Search products..."
