+++ frontend/src/context/SearchContext.tsx
